6. python-esport-llm/requirements.txt
discord.py==2.3.2
audioop-lts==0.2.2; python_version>="3.13"